add goobi integration class and refactor
